+++ Hyperar.HattrickUltimate.BusinessObjects/Hattrick/TeamDetails/SupporterTeam.cs
@@ -16,12 +16,12 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.TeamDetails
         /// <summary>
         /// Gets or sets the League Id.
         /// </summary>
-        public long LeagueId { get; set; }
+        public uint LeagueId { get; set; }
 
         /// <summary>
         /// Gets or sets the LeagueLevelUnit Id.
         /// </summary>
-        public long LeagueLevelUnitId { get; set; }
+        public uint LeagueLevelUnitId { get; set; }
 
         /// <summary>
         /// Gets or sets the LeagueLevelUnit Name.
@@ -41,7 +41,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.TeamDetails
         /// <summary>
         /// Gets or sets the Team Id.
         /// </summary>
-        public long TeamId { get; set; }
+        public uint TeamId { get; set; }
 
         /// <summary>
         /// Gets or sets the Team Name.
@@ -51,7 +51,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.TeamDetails
         /// <summary>
         /// Gets or sets the User Id.
         /// </summary>
-        public long UserId { get; set; }
+        public uint UserId { get; set; }
 
         #endregion Public Properties
     }
